Put dot module in subfolder with separate file for dot_parser
